Move some non-template functions into a new file.
This will let us split up grid_tools.cc with SPLIT_INSTANTIATIONS_COUNT
in the next step.
