+++ source/grid/CMakeLists.txt
@@ -37,6 +37,7 @@ SET(_separate_src
   grid_generator.cc
   grid_tools.cc
   grid_tools_cache.cc
+  grid_tools_nontemplates.cc
   grid_tools_dof_handlers.cc
   tria.cc
   )
